fixed all errors thrown due to renaming files
