+++ index.js
@@ -1,6 +1,7 @@
 const { ApolloServer } = require('apollo-server')
 const mongoose = require('mongoose')
 const schema = require('./modules')
+const authSchema = require('./modules/auth/typeDefs')
 // const typeDefs = require("./graphql/typeDefs")
 // const resolvers = require("./graphql/resolvers")
 const { MONGODB, PORT } = require("./config/constants")
