+++ modules/index.js
@@ -1,7 +1,7 @@
 const { makeExecutableSchemaFromModules } = require('../utils/modules')
 
 const auth = require('./auth/typeDefs')
-const users = require('./users')
+const users = require('./users/typeDefs')
 
 module.exports = makeExecutableSchemaFromModules({
   modules: [
